263
I'm trying to think about how to write a test for the race condition,
and how to fix it. One thing that's been hard is even remembering where
it lies. It's not between wiping the watch and sleeping on it; that's
innocuous because the sleep would just immediately wake up. No, the race
condition lies between the empty check and the wipe.

For the innocuous race we could just create an atomic wipe-and-sleep.
But the more serious race requires a lock.

If we need a lock anyway, is there any reason to have two watch
variables?

I'm going to preserve these alternative functions in the code.
Alternatives will only ever be called from other alteratives or tests.