319 - ack, forgot to handle blocks when tangling
Will the 'lightweight tools' really be all that useable if we encourage
people to layer them one atop another and track precisely what inputs
each can accept? Something to keep an eye on.

In the meanwhile, we have a new (but very unrealistic) example
demonstrating the tangling directives.

There's still a big constraint on ordering: before/after clauses have to
come before functions that need them.