332 - example function with multiple clauses
I'm still figuring out the idioms, and it doesn't seem like there's an
easy way to avoid redundant allocations and arg reading.

At least for generic functions we want to optimize for each call-site,
we can create subsidiary labels to jump to.

Also, this is the first genuine use for random access to a function's
arguments.