464 - reply broken in channel 'write'
How did this work until now? The reply was being treated as a label, and
if 'write' returned nothing it would still work fine because the output
is already present, and a missing 'reply' leaves oargs as-is.

Should we do something to catch this? Perhaps we should track args
modified and check that there are oargs for them. But that seems quite
heavyweight.. Maybe we should clear oargs when missing a 'reply'?