605 - example program for line-oriented input
We need the hack of printing characters typed to screen right after we
see them. But only when it's the real screen. And there's no way to
write a test for that because it explicitly shouldn't happen for fake
terminals :(

Never mind, we'll be able to test it when we provide some mechanism for
suppressing print. The equivalent of 'stty -echo'.