674 - eliminate the other label
Jumps forward are harder than jumps back. You can't create a
continuation until you get to it. And we don't have a way to return
things from a continuation yet. That's a flaw. Maybe we need a
'reply-from' operator.

But for now, this small bit of code we can just inline and duplicate.