745 - first test of expand functionality
Expanding trace items, of course, doesn't work on anything but the last
item. But we've now caught automatic tests up to the state of
development.

These tests are hugely expensive. Previous .t files run in ~30s. Already
the trace browser tests clock in at 2 minutes. And we're just getting
started. Time to explore reimplementing the mu VM in C.