1030 - better test isolation for break/loop
I've been resetting all recipes after every single test, but the arc
version has shown that this gets slow all too quickly. And the longer I
wait to fix it the harder it gets to fix. Already, boy, were this and
the next couple of commits hard to track down.