1441 - give up on unit tests for 'random'
I'd been hoping that I could simply pass in the previously-returned
number to srand() to generate the next one in the series. But looks like
rand() is more stateful than that.

Another weirdness: I put in 'round' in the same layer because of the
vague idea that it would help generate random integers. But that's all
we get anyway.