1458
While pushing out color support in fake screens I realized I've been
complecting the special-case of a special-case to transform
literal-string arguments for 'new'. As a result I hadn't been catching
bad habits like giving its arg the wrong type. Now we have cleaner
separation of the two variants of 'new', a few more checks, and better
error messages when we mis-call it.

Aside: I've added a third goto target. Sliding into spaghetti? Keep an
eye on it.

This goto might become a common pattern: a layer hooking into a previous
one to prevent it from happening. In this case new on literal-strings
prevents the transform for new from triggering.