1551 - initial sketch of layout
Two columns: edit code on left and run code on right.
In both columns, the action is at the top of the screen, and stuff
gradually flows downward unless pulled back up.
Stuff you run on the right side will eventually become reproducible test
cases that runs constantly on every change.

We also have a tentative signature for the 'edit' routine: it'll take an
input string and return an output string when it finishes.

More columns expand to the right. Number of columns shown will remain
fixed at any time depending on screen width. But for now we support
exactly two columns.