1564 - a better way to support string literals
Our new heuristic is: all string literals are the same. If they contain
newline before non-whitespace, we scan for comments assuming there might
be code inside:

  foofoofoo [
    ...  # ']' inside comment ignored
  ]

If they contain non-whitespace first, then we ignore comments assuming
it's just a regular string:

  foofoofoo [abc#def]  # valid string literal

The big hole in this approach:

  foofoofoo [ # what about comments here containing ']'?
    ... # abc
  ]

Currently this reads as a 'code comment' and terminates before the
newline or '?' and will probably trigger errors down the line.
Temporary workaround: don't start code strings with a comment on the
same line as the '['. Eventually we'll tighten up the logic.

We're still not using the new heuristic in scenarios, but that's up
next.