1658 - disallow cursor from being 'outside' text
It can at most be just after the text of a line. If I click further
right, adjust it so it's at the end. If I click below, adjust it so it's
at the end.

Still not perfect. We aren't testing the value of the before-cursor
pointer, and it's actually still wrong when clicking below all the text
in the editor.