1721 - hide warnings inside interactive routines
We will need many other forms of isolation for these. For starters we're
going to have to replace most asserts with warnings that can be traced
so that the environment doesn't crash because of illegal code typed into
it.

New test is still failing. Just getting it to fail right was hard
enough.