1826 - edit: start carefully showing all errors
Eventually we might be able to get rid of die entirely.

This is just a preliminary stab at a random error. In the process I ran
into two issues that have impeded debugging before:

a) Naming conflicts within scenarios are a real no-no. I need to warn on
them, but the rules are getting complicated:
  Always print warnings on redefine
  But not in interactive mode
  Or in scenarios checking warning behavior
  Unless the scenario recipe itself is overridden

b) Now that we've added collect_layers and a long time can go between
traces, debugging is a minefield because trace lines don't print to
screen immediately after they're created. Need to do something about
that. Maybe explicitly trigger collection by tracing '\n' or something.

These are the next two items on my todo list.