2046
Redo some of the work of commits 2025-2032. We still need to create
special-cases for each branch, but we'll now always update cursor in a
single place. We'll also have handle-keyboard-event tell its caller
whether it already took care of rendering or not. That way the caller
only falls back to painting the entire screen when necessary. No more
double-painting.