2052
If I try to run a single test and it triggers an error the trace gets
saved in the current directory, as if I was trying to log an interactive
run. Then when I try to rerun the test the trace tries to load as mu
code, and hilarity ensues. Just log interactive runs in .traces/ as well.