start using the new check
I wasn't seeing errors because I wasn't using /contained-in in products
yet. But it seems to work fine even after.

One reason this isn't an invasive change is that it's opt-in. Most of
the time it isn't triggered. You have to add the /contained-in check to
trigger a check. But that's just like regular const checks in other
languages: if you don't specify immutability you get no checks.