2585 - label sandboxes with a number
It also seems useful that the number maps to the name of the file the
sandbox is saved in. However this mapping is currently a happy accident
and not actually tested.

I'm starting to switch gears and help make the editor useable with
many many sandboxes. This is just the first step of several.