2608 - fix-up tests in sandbox/ app
When I first forked it from the edit/ app, I wasn't sure how to deal
with changing the recipe side when the only way the program accesses it
is with the untestable 'restore' hack. Now we introduce a little hook
into event-loop and pass in any updated recipe side directly.

In the process I've cleaned up several minor stylistic inconsistencies
between edit/ and sandbox/ apps.