2640 - fix some tests
Thanks Caleb Couch for finding this. These tests were breaking only when
some other code somewhere has already triggered the specialization of
'append' for numbers.

How to test this? Perhaps the right fix is to warn when character
variants are used with literals. Or only when there's also a generic
variant that could conceivably be specialized for numbers.