2674
Make it explicit that I don't currently intend to further replace type
ingredients inside replacements.

That might well make sense, but I've not thought up any use cases or
tests for it, and it seems excessive at the moment. And when I find bugs
I keep wondering if the repeated replace might be the problem. No
longer.