2798 - experiment: split channels into two ends
This way when you pass one end to a function or routine, you can
implicitly give it the right to either read or write the channel, but
not both.

The cost: code gets more convoluted, names get more convoluted. You can
see this in particular in the test for buffer-lines. Let's see how it
goes..