2806 - bugfix: cleaning up in 'reload'
This brings back some of the complexity I thought I'd gotten rid of in
2799.

The regression brought home the point that I'd forgotten to write tests
for those bits. Written now.

It also brought home the point that our cleanup in 'reload' has always
been hacky and incomplete.

It's also ugly that those tests in the sandbox/ and edit/ apps need
changing (particularly when the test is about how the output doesn't
change).