2870 - fix the final long-standing failing test
The solution for avoiding deadlock is for routines to close channels
before they exit. So that's good.

Once I implemented 'close', I also found and fixed 2 unrelated bugs in
chessboard.mu:
  a) one long-missed and long-masked case of forgetting to store
  character literals in character variables
  b) one typo in translating get-address to put
So that's good.

What's not so good: in the process of fixing this I've found three
unrelated bugs (marked 'BUG:' in the changes). All three have
workarounds, so existing tests pass for now. But they are my top
priority next.