3010
insert_container is getting pretty gnarly. It's spread across two layers
(containers and shape-shifting containers), and since it has to deal
with extending existing containers it's coiled in on itself,
constantly reading and writing the Type table.

Maybe I should uncoil the case of extending a container and make it
separate from defining a new container.