3038 - track down a long-standing bug
In some rare situations the editor would join a line with the next when
it should simply wrap to the next screen row. Thanks Caleb and Ella
Couch for finally running into a situation that was easy to reproduce.

The scenario diffs are misleading on this commit. I had to:

a) delete the obsolete 'editor-wraps-cursor-after-inserting-characters'
because it was written back when a line just large enough to fit in a
single line would not wrap:

   |     |  <-- screen boundary
    abcde

These days it will wrap after making room for the wrap indicator:

   |     |  <-- screen boundary
    abcd↩
    e

b) rename editor-wraps-cursor-after-inserting-characters-2 to
editor-wraps-cursor-after-inserting-characters-in-middle-of-line

c) create a new scenario demonstrating the bug:
editor-wraps-cursor-after-inserting-characters-at-end-of-line