3088
Gaining confidence now in the fix of 3086. Basically I forgot that
size_of doesn't canonize its argument (because it needs to run at
transform-time, but canonize cares about the value of the reagent for
arrays, and value is only available at run-time). So I was always
returning the size of the address (1) as the size of its payload.