3163
Experimental: kinda support $print in console mode.

It's not perfect and probably will never be, because 'cout' buffers
differently from termbox primitives, which can cause console-aware
newlines to show up before other (console-oblivious) prints, like in
this example program:

  def main [
    open-console
    $print [abc], 10/newline
    $print [def], 10/newline
    wait-for-some-interaction
    close-console
  ]

And then there's the problem that there's no way for cout to update
Display_column. So mixing $print and print will be confusing.

Perhaps we should just not mess with Display_* variables inside $print?
But then we'll only ever be able to see a single line of $print at a
time.